+++ components/StarSearch/StarSearchChat.tsx
@@ -64,7 +64,7 @@ async function updateComponentRegistry(name: string) {
 }
 
 type StarSearchChatProps = {
-  userId: number | undefined;
+  userId: number | null | undefined;
   sharedPrompt?: string | null;
   sharedChatId?: string | null;
   bearerToken: string | undefined | null;
@@ -76,6 +76,7 @@ type StarSearchChatProps = {
   onClose?: () => void;
   baseApiStarSearchUrl?: URL;
   sharingEnabled?: boolean;
+  isWorkspace?: boolean;
 };
 
 export function StarSearchChat({
@@ -91,6 +92,7 @@ export function StarSearchChat({
   baseApiStarSearchUrl = DEFAULT_STAR_SEARCH_API_BASE_URL,
   sharingEnabled = true,
   showTopNavigation = false,
+  isWorkspace = false,
 }: StarSearchChatProps) {
   const [starSearchState, setStarSearchState] = useState<"initial" | "chat">("initial");
   const [chat, setChat] = useState<StarSearchChatMessage[]>([]);
@@ -98,7 +100,7 @@ export function StarSearchChat({
   const [isRunning, setIsRunning] = useState(false);
   const [showSuggestions, setShowSuggestions] = useState(false);
   const [ranOnce, setRanOnce] = useState(false);
-  const { feedback, prompt } = useStarSearchFeedback();
+  const { feedback, prompt } = useStarSearchFeedback(isWorkspace);
   const { toast } = useToast();
   const [loginModalOpen, setLoginModalOpen] = useState(false);
   const [checkAuth, setCheckAuth] = useState(false);