+++ CHANGELOG.md
@@ -5,6 +5,49 @@
 
 > All notable changes to this project will be documented in this file
 
+## [2.46.0-beta.6](https://github.com/open-sauced/app/compare/v2.46.0-beta.5...v2.46.0-beta.6) (2024-07-18)
+
+
+### 🍕 Features
+
+* added posthog analytics for StarSearch for Workspaces ([#3748](https://github.com/open-sauced/app/issues/3748)) ([72d7c23](https://github.com/open-sauced/app/commit/72d7c2335867b8cafb2ac79189ad312ef2f7569b))
+
+## [2.46.0-beta.5](https://github.com/open-sauced/app/compare/v2.46.0-beta.4...v2.46.0-beta.5) (2024-07-18)
+
+
+### 🐛 Bug Fixes
+
+* now StarSearch for workspaces only works for editors ([#3733](https://github.com/open-sauced/app/issues/3733)) ([4aeb220](https://github.com/open-sauced/app/commit/4aeb220dec334e37629537b61d5959e5b1cb73f2))
+
+## [2.46.0-beta.4](https://github.com/open-sauced/app/compare/v2.46.0-beta.3...v2.46.0-beta.4) (2024-07-18)
+
+
+### 🍕 Features
+
+* Tracking Repo Pages (YOLO Chart) ([#3736](https://github.com/open-sauced/app/issues/3736)) ([5edfe8a](https://github.com/open-sauced/app/commit/5edfe8ac5cbefd04ca36e0fbbda85bdc330c5f1d))
+
+## [2.46.0-beta.3](https://github.com/open-sauced/app/compare/v2.46.0-beta.2...v2.46.0-beta.3) (2024-07-18)
+
+
+### 🍕 Features
+
+* Tracking Repo Pages (Miscellaneous) ([#3739](https://github.com/open-sauced/app/issues/3739)) ([5530beb](https://github.com/open-sauced/app/commit/5530beb1868d925c67a9390e8b924289d36d6e58))
+
+## [2.46.0-beta.2](https://github.com/open-sauced/app/compare/v2.46.0-beta.1...v2.46.0-beta.2) (2024-07-17)
+
+
+### 🍕 Features
+
+* removed StarSearch feature flagging ([#3730](https://github.com/open-sauced/app/issues/3730)) ([701a95a](https://github.com/open-sauced/app/commit/701a95a3294c81ef878b4ed6dee6db380bdf9c47))
+* Tracking Repo Pages (Lottery Factor Chart) ([#3734](https://github.com/open-sauced/app/issues/3734)) ([be5c72e](https://github.com/open-sauced/app/commit/be5c72e203c9406128fe30b90d355584e5de1354))
+
+## [2.46.0-beta.1](https://github.com/open-sauced/app/compare/v2.45.0...v2.46.0-beta.1) (2024-07-17)
+
+
+### 🍕 Features
+
+* Track Repo Pages (Header) ([#3728](https://github.com/open-sauced/app/issues/3728)) ([2dede76](https://github.com/open-sauced/app/commit/2dede76509a7a489c63829e644a8c9cb219be36a))
+
 ## [2.45.0](https://github.com/open-sauced/app/compare/v2.44.0...v2.45.0) (2024-07-16)